Updated the structure to follow descussed future practices
