+++ packages/web-app/src/components/primitives/titles/MenuTitle.styles.tsx
@@ -1,5 +1,5 @@
 // Theme
-import { SaladTheme } from '../../SaladTheme'
+import { SaladTheme } from '../../../SaladTheme'
 
 export const styles = (theme: SaladTheme) => ({
   menuTitle: {
@@ -17,7 +17,5 @@ export const styles = (theme: SaladTheme) => ({
         color: `${theme.lightGreen} !important`
       }
     },
-    
-    
   }
 })
