+++ packages/web-app/src/modules/settings-views/battery-saver-views/components/BatterySaver.tsx
@@ -4,33 +4,25 @@ import React, { Component } from 'react'
 import { styles } from './BatterySaver.styles'
 
 // UI
-import {
-  VeggieName,
-  CondensedHeader,
-  Body,
-} from '../../../../ui'
+import { VeggieName, CondensedHeader, AppBody } from '../../../../components'
 
 // Packages
 import withStyles, { WithStyles } from 'react-jss'
 // import classnames from 'classnames'
 
-interface Props extends WithStyles<typeof styles> {
-
-}
+interface Props extends WithStyles<typeof styles> {}
 
 class _BatterySaver extends Component<Props> {
   render() {
     return (
       <>
-        <header>
-          <VeggieName value="Coming Soon" />
-          <CondensedHeader value="Battery Saver" />
-        </header>
-        <main>
-          <Body>
-            Salad won’t chop if your laptop isn’t connected to a power source.
-          </Body>
-        </main>
+        <div className="header">
+          <VeggieName>Coming Soon</VeggieName>
+          <CondensedHeader>Battery Saver</CondensedHeader>
+        </div>
+        <div className="main">
+          <AppBody>Salad won’t chop if your laptop isn’t connected to a power source.</AppBody>
+        </div>
       </>
     )
   }
