--- packages/web-app/src/ui/body/Body.styles.tsx
@@ -1,10 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  body: {
-    fontFamily: theme.fontGroteskBook19,
-    fontSize: theme.small,
-    lineHeight: theme.medium,
-  }
-})
