--- packages/web-app/src/ui/headers/CondensedHeader.styles.tsx
@@ -1,14 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  condensedHeader: {
-    color: theme.darkBlue,
-    fontFamily: theme.fontGroteskLight09,
-    fontSize: theme.xLarge,
-    lineHeight: theme.xLarge,
-    fontWeight: 'normal',
-    margin: 0,
-    padding: 0,
-  }
-})
