--- packages/web-app/src/ui/index.ts
@@ -1,17 +1,0 @@
-// Body
-export * from './body/Body'
-
-// Headers
-export * from './headers/CondensedHeader'
-
-// Lists
-export * from './lists/LinkListUnstyled'
-export * from './lists/ListUnstyled'
-
-// Percentages
-export * from './percentages/Percentage'
-
-// Titles
-export * from './titles/MenuTitle'
-export * from './titles/Username'
-export * from './titles/VeggieName'
