--- packages/web-app/src/ui/lists/LinkListUnstyled.styles.tsx
@@ -1,11 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  linkListUnstyled: {
-    margin: 0,
-    padding: 0,
-    listStyle: 'none',
-  },
-  linkListItem: {},
-})
