--- packages/web-app/src/ui/lists/ListUnstyled.styles.tsx
@@ -1,9 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  listUnstyled: {
-    paddingLeft: 0,
-    listStyle: 'none',
-  }
-})
