--- packages/web-app/src/ui/percentages/Percentage.styles.tsx
@@ -1,11 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  percentage: {
-    color: theme.orange,
-    fontFamily: theme.fontGroteskLight05,
-    fontSize: theme.xxLarge,
-    lineHeight: theme.xxLarge,
-  }
-})
