--- packages/web-app/src/ui/titles/Username.styles.tsx
@@ -1,11 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  username: {
-    color: theme.green,
-    fontFamily: theme.fontGroteskLight25,
-    fontSize: theme.mediumLarge,
-    lineHeight: theme.mediumLarge,
-  }
-})
