--- packages/web-app/src/ui/titles/VeggieName.styles.tsx
@@ -1,14 +1,0 @@
-// Theme
-import { SaladTheme } from '../../SaladTheme'
-
-export const styles = (theme: SaladTheme) => ({
-  veggieName: {
-    fontFamily: theme.fontGroteskLight25,
-    fontSize: theme.xSmall,
-    lineHeight: theme.xSmall,
-    letterSpacing: 1,
-    textTransform: 'uppercase',
-    margin: 0,
-    padding: 0,
-  }
-})
